requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
apscheduler>=3.10.4
pydantic-settings>=2.2.0
pycryptodome>=3.20.0
python-multipart>=0.0.9
lark-oapi>=1.2.0
wechatpy>=1.8.18
xmltodict>=0.13.0
wecom-aibot-sdk-python>=0.1.6
